+++ src/main/java/app/controller/StartscreenController.java
@@ -1,5 +1,6 @@
 package app.controller;
 
+import app.ApplicationState;
 import app.elements.IconPicker;
 import app.model.OpenLobbiesModel;
 import app.model.StartScreenModel;
@@ -37,14 +38,17 @@ public class StartscreenController extends Controller implements StartScreenMode
 
   @FXML
   private void switchToLobbyBrowser() {
-    try {
-      if (username.getText().equals("")) {
-        displayError(USERNAME_ERROR);
-      } else {
-        model.register(username.getText());
+    if (ApplicationState.getInstance().getClient() == null) {
+      try {
+        ApplicationState.getInstance().newClient();
+        if (username.getText().equals("")) {
+          displayError(USERNAME_ERROR);
+        } else {
+          model.register(username.getText());
+        }
+      } catch (IOException e) {
+        displayError(e.getMessage());
       }
-    } catch (Exception e) {
-      e.printStackTrace();
     }
   }
 
@@ -69,9 +73,13 @@ public class StartscreenController extends Controller implements StartScreenMode
   public void registered() {
     try {
       new OpenLobbiesController(stage, new OpenLobbiesModel()).show();
-      model.setObserver(null);
     } catch (IOException e) {
       e.printStackTrace();
     }
+  }
+
+  @Override
+  public void receivedError(String message) {
+    displayError(message);
   }
 }