+++ src/main/java/app/model/CreateModel.java
@@ -3,13 +3,14 @@ package app.model;
 import app.ApplicationState;
 import app.IconManager;
 import client.Client;
+import client.ErrorObserver;
 import client.LobbyObserver;
 import javafx.application.Platform;
 import protocol.LobbyData;
 import protocol.RaceData;
 
 /** Model for CreateView. */
-public class CreateModel implements LobbyObserver {
+public class CreateModel implements LobbyObserver, ErrorObserver {
 
   private CreateModelObserver observer;
 
@@ -31,7 +32,7 @@ public class CreateModel implements LobbyObserver {
     Client client = ApplicationState.getInstance().getClient();
     String userId = ApplicationState.getInstance().getUserId();
     String iconId = IconManager.getSelectedIcon().getId();
-    client.subscribeLobbyUpdates(this);
+    subscribe();
     client.newLobby(userId, name, iconId);
   }
 
@@ -45,6 +46,25 @@ public class CreateModel implements LobbyObserver {
     if (observer != null) {
       Platform.runLater(() -> observer.joinedLobby(lobby));
     }
-    ApplicationState.getInstance().getClient().unsubscribeLobbyUpdates(this);
+    unsubscribe();
+  }
+
+  @Override
+  public void receivedError(String message) {
+    if (observer != null) {
+      Platform.runLater(() -> observer.receivedError(message));
+    }
+  }
+
+  private void subscribe() {
+    Client client = ApplicationState.getInstance().getClient();
+    client.subscribeErrors(this);
+    client.subscribeLobbyUpdates(this);
+  }
+
+  private void unsubscribe() {
+    Client client = ApplicationState.getInstance().getClient();
+    client.unsubscribeErrors(this);
+    client.unsubscribeLobbyUpdates(this);
   }
 }