+++ src/main/java/app/model/StartScreenModel.java
@@ -3,12 +3,13 @@ package app.model;
 import app.ApplicationState;
 import client.Client;
 import client.ClientObserver;
+import client.ErrorObserver;
 import java.util.List;
 import javafx.application.Platform;
 import protocol.LobbyData;
 
 /** Model for StartScreen View. */
-public class StartScreenModel implements ClientObserver {
+public class StartScreenModel implements ClientObserver, ErrorObserver {
 
   private StartScreenModelObserver observer;
 
@@ -19,7 +20,7 @@ public class StartScreenModel implements ClientObserver {
    */
   public void register(String name) {
     Client client = ApplicationState.getInstance().getClient();
-    client.subscribe(this);
+    subscribe();
     client.registerUser(name);
   }
 
@@ -34,15 +35,34 @@ public class StartScreenModel implements ClientObserver {
 
   @Override
   public void registered(String userId) {
+    unsubscribe();
     ApplicationState.getInstance().setUserId(userId);
     if (observer != null) {
       Platform.runLater(() -> observer.registered());
     }
-    ApplicationState.getInstance().getClient().unsubscribe(this);
   }
 
   @Override
   public void receivedOpenLobbies(List<LobbyData> lobbies) {
     // TODO: remove unused functions from interface
+  }
+
+  @Override
+  public void receivedError(String message) {
+    if (observer != null) {
+      Platform.runLater(() -> observer.receivedError(message));
+    }
+  }
+
+  private void subscribe() {
+    Client client = ApplicationState.getInstance().getClient();
+    client.subscribeErrors(this);
+    client.subscribe(this);
+  }
+
+  private void unsubscribe() {
+    Client client = ApplicationState.getInstance().getClient();
+    client.unsubscribeErrors(this);
+    client.unsubscribe(this);
   }
 }