+++ src/main/java/typeracer/GamePhase.java
@@ -1,4 +1,4 @@
-package model;
+package typeracer;
 
 /** Enum class for game phases. */
 public enum GamePhase {